Cleanup

+++ distributions/archlinux/install/apps/docker.sh
@@ -4,7 +4,9 @@ yay -S --noconfirm --needed docker docker-compose docker-buildx
 
 # Limit log size to avoid running out of disk
 sudo mkdir -p /etc/docker
-echo '{"log-driver":"json-file","log-opts":{"max-size":"10m","max-file":"5"}}' | sudo tee /etc/docker/daemon.json
+if [ ! -f /etc/docker/daemon.json ]; then
+  echo '{"log-driver":"json-file","log-opts":{"max-size":"10m","max-file":"5"}}' | sudo tee /etc/docker/daemon.json
+fi
 
 # Start Docker automatically
 sudo systemctl enable docker
@@ -14,10 +16,13 @@ sudo usermod -aG docker ${USER}
 
 # Prevent Docker from preventing boot for network-online.target
 sudo mkdir -p /etc/systemd/system/docker.service.d
-sudo tee /etc/systemd/system/docker.service.d/no-block-boot.conf <<'EOF'
+# Apply if not exist
+if [ ! -f /etc/systemd/system/docker.service.d/no-block-boot.conf ]; then
+  sudo tee /etc/systemd/system/docker.service.d/no-block-boot.conf <<'EOF'
 [Unit]
 DefaultDependencies=no
 EOF
+fi
 
 sudo systemctl daemon-reload
